// brokerchain-1.0.0/src/library/file-system/file-ls/type.ts
export interface Input {
    dir: string;
    recursive: boolean;
}

export interface OutputEmpty {
    file_list: { name: string; fullname: string }[];
}

export interface OutputOk {
    file_list: { name: string; fullname: string }[];
}

export interface OutputFail {}

export interface Callback<R> {
    empty: (output: OutputEmpty) => R;
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}
// ...
export type Output = OutputOk;